requirements.txt
slack-bolt==1.18.1
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0 